+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Change Chiplet Fields to Public (#1629).
 - Added to the `Assembler` the ability to vendor a compiled library.
 - Introduce `BusDebugger` to facilitate debugging buses (#1664)
+- [BREAKING] Introduced `HORNERBASE`, `HORNEREXT` and removed `RCOMBBASE` (#1656).
 
 
 ## 0.12.0 (2025-01-22)